+++ Covenant/Models/Launchers/CscriptLauncher.cs
@@ -12,7 +12,7 @@ using Covenant.Core;
 
 namespace Covenant.Models.Launchers
 {
-    public class CscriptLauncher : ScriptletLauncher, ILoggable
+    public class CscriptLauncher : ScriptletLauncher
     {
         public CscriptLauncher()
         {
@@ -42,10 +42,5 @@ namespace Covenant.Models.Launchers
             }
             else { return ""; }
         }
-        //public void ToLog(LogAction action, LogLevel level)
-        //{
-            // CScriptLauncher|Action|ID|Name|LauncherString|ScriptType|OutputKind|CompressStager
-        //    Common.logger.Log(level, $"{this.Type}|{action}|{this.Id}|{this.Name}|{this.LauncherString}|{this.ScriptType}|{this.OutputKind}|{this.CompressStager}");
-        //}
     }
 }