+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added ability to specify custom admin port number
 - Added Covenant trace logging
+- Added ServiceBinary launcher
 
 ### Changed
 - Disallow Administrators from changing other user's passwords